Split off GameCommand enum to its own header as well
