+++ src/openrct2-ui/windows/ServerStart.cpp
@@ -14,6 +14,7 @@
     #include <openrct2-ui/interface/Widget.h>
     #include <openrct2-ui/windows/Windows.h>
     #include <openrct2/Context.h>
+    #include <openrct2/Game.h>
     #include <openrct2/ParkImporter.h>
     #include <openrct2/config/Config.h>
     #include <openrct2/core/String.hpp>
