+++ src/openrct2/Game.h
@@ -35,98 +35,6 @@ namespace OpenRCT2
     constexpr float kGameMaxTimeScale = 5.0f;
 } // namespace OpenRCT2
 
-struct ParkLoadResult;
-
-enum class GameCommand : int32_t
-{
-    SetRideAppearance,
-    SetLandHeight,
-    TogglePause,
-    PlaceTrack,
-    RemoveTrack,
-    LoadOrQuit,
-    CreateRide,
-    DemolishRide,
-    SetRideStatus,
-    SetRideVehicles,
-    SetRideName,
-    SetRideSetting,
-    PlaceRideEntranceOrExit,
-    RemoveRideEntranceOrExit,
-    RemoveScenery,
-    PlaceScenery,
-    SetWaterHeight,
-    PlacePath,
-    PlacePathLayout,
-    RemovePath,
-    ChangeSurfaceStyle,
-    SetRidePrice,
-    SetGuestName,
-    SetStaffName,
-    RaiseLand,
-    LowerLand,
-    EditLandSmooth,
-    RaiseWater,
-    LowerWater,
-    SetBrakesSpeed,
-    HireNewStaffMember,
-    SetStaffPatrol,
-    FireStaffMember,
-    SetStaffOrders,
-    SetParkName,
-    SetParkOpen,
-    BuyLandRights,
-    PlaceParkEntrance,
-    RemoveParkEntrance,
-    SetMazeTrack,
-    SetParkEntranceFee,
-    SetStaffColour,
-    PlaceWall,
-    RemoveWall,
-    PlaceLargeScenery,
-    RemoveLargeScenery,
-    SetCurrentLoan,
-    SetResearchFunding,
-    PlaceTrackDesign,
-    StartMarketingCampaign,
-    PlaceMazeDesign,
-    PlaceBanner,
-    RemoveBanner,
-    SetSceneryColour,
-    SetWallColour,
-    SetLargeSceneryColour,
-    SetBannerColour,
-    SetLandOwnership,
-    ClearScenery,
-    SetBannerName,
-    SetSignName,
-    SetBannerStyle,
-    SetSignStyle,
-    SetPlayerGroup,
-    ModifyGroups,
-    KickPlayer,
-    Cheat,
-    PickupGuest,
-    PickupStaff,
-    BalloonPress,
-    ModifyTile,
-    EditScenarioOptions,
-    PlacePeepSpawn,
-    SetColourScheme,
-    SetStaffCostume,
-    PlaceFootpathAddition,
-    RemoveFootpathAddition,
-    GuestSetFlags,
-    SetDate,
-    Custom,
-    ChangeMapSize,
-    FreezeRideRating,
-    SetGameSpeed,
-    SetRestrictedScenery,
-    setRideVisibility,
-    Count,
-};
-
 enum
 {
     GAME_PAUSED_NORMAL = 1 << 0,
