+++ src/openrct2/actions/GameAction.hpp
@@ -9,11 +9,11 @@
 
 #pragma once
 
-#include "../Game.h"
 #include "../core/DataSerialiser.h"
 #include "../core/Identifier.hpp"
 #include "GameActionParameterVisitor.h"
 #include "GameActionResult.h"
+#include "GameCommand.h"
 
 #include <functional>
 #include <memory>
