+++ src/openrct2/actions/GameSetSpeedAction.cpp
@@ -10,6 +10,7 @@
 #include "GameSetSpeedAction.h"
 
 #include "../Diagnostic.h"
+#include "../Game.h"
 #include "../config/Config.h"
 #include "../ui/WindowManager.h"
