+++ src/openrct2/actions/LoadOrQuitAction.cpp
@@ -10,6 +10,7 @@
 #include "LoadOrQuitAction.h"
 
 #include "../Context.h"
+#include "../Game.h"
 #include "../OpenRCT2.h"
 #include "../ui/WindowManager.h"
