+++ src/openrct2/network/NetworkAction.h
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include "../Game.h"
+#include "../actions/GameCommand.h"
 #include "../localisation/StringIdType.h"
 
 #include <array>
