+++ src/openrct2/network/NetworkPlayer.h
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include "../Game.h"
+#include "../actions/GameCommand.h"
 #include "../core/Money.hpp"
 #include "../world/Location.hpp"
